# internal/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_krkn-chaos_krkn-operator_api_v1alpha1.GraphScenarioNode:
    properties:
      _comment:
        description: |-
          Comment is an optional comment describing the scenario
          +optional
        type: string
      cloudCredentialRef:
        description: |-
          CloudCredentialRef overrides the graph-level cloud credential for this specific node
          +optional
        type: string
      depends_on:
        description: |-
          DependsOn is the node ID that this scenario depends on (parent in the graph)
          +optional
        type: string
      env:
        additionalProperties:
          type: string
        description: |-
          Env is a map of environment variables for the scenario
          +optional
        type: object
      image:
        description: |-
          Image is accepted for backwards compatibility but is never used to select
          a pod image; the operator resolves images from Scenario.
          +optional
        type: string
      name:
        description: |-
          Name is the legacy scenario identity field. When Scenario is absent, the
          API translates it into a public ScenarioReference.
          +optional
        type: string
      registryName:
        description: |-
          RegistryName is the legacy private registry identity field. Image remains
          ignored even when this field is present.
          +optional
        type: string
      scenario:
        allOf:
        - $ref: '#/definitions/github_com_krkn-chaos_krkn-operator_api_v1alpha1.ScenarioReference'
        description: |-
          Scenario identifies the scenario and registry to resolve for this node.
          Metadata-only nodes (whose IDs start with _) may omit this field.
          +optional
      volumes:
        additionalProperties:
          type: string
        description: |-
          Volumes is a map of volume mounts for the scenario
          +optional
        type: object
    type: object
  github_com_krkn-chaos_krkn-operator_api_v1alpha1.ScenarioReference:
    properties:
      name:
        description: Name is the scenario tag/name to resolve.
        type: string
      private:
        description: |-
          Private selects a saved private registry when true, or krknctl's public
          Quay provider when false. A pointer makes the field mandatory on input.
        type: boolean
      registryName:
        description: RegistryName identifies the saved private registry when Private
          is true.
        type: string
    type: object
...
  github_com_krkn-chaos_krkn-operator_pkg_files.FileReference:
    properties:
      fileId:
        description: FileID is the UUID of the file to mount
        type: string
      mountPath:
        description: MountPath is the absolute path where the file should be mounted
        type: string
    type: object
...
  internal_api.ClusterResiliencyScoreResponse:
    properties:
      clusterName:
        type: string
      message:
        type: string
      score:
        type: number
      status:
        type: string
    type: object
...
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  internal_api.FileMount:
    properties:
      content:
        description: Content is the base64-encoded file content
        type: string
      mountPath:
        description: MountPath is the absolute path where the file should be mounted
        type: string
      name:
        description: Name is the file name
        type: string
    type: object
...
  internal_api.GraphRunCreateRequest:
    properties:
      categories:
        description: Categories are visible category names to associate with the run
          when it is created.
        items:
          type: string
        type: array
      cloudCredentialRef:
        description: CloudCredentialRef is the default cloud credential for all nodes
          (optional)
        type: string
      graph:
        additionalProperties:
          $ref: '#/definitions/github_com_krkn-chaos_krkn-operator_api_v1alpha1.GraphScenarioNode'
        description: Graph is the dependency graph of scenarios to execute
        type: object
      targetClusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: TargetClusters is a map of provider-name to list of cluster names
        type: object
      targetRequestId:
        description: TargetRequestID is the reference to the KrknTargetRequest CR
        type: string
    type: object
...
  internal_api.PaginationMeta:
    properties:
      limit:
        description: Limit is the number of items per page, 0 if unpaginated
        type: integer
      page:
        description: Page is the current page number (1-based), 0 if unpaginated
        type: integer
      total:
        description: Total is the total number of items matching the query
        type: integer
      totalPages:
        description: TotalPages is the total number of pages
        type: integer
    type: object
...
  internal_api.ScenarioRunListItem:
    properties:
      createdAt:
        description: CreatedAt is the creation timestamp
        type: string
      customRunName:
        description: CustomRunName is the user-provided label for this run
        type: string
      failedJobs:
        description: FailedJobs is the number of failed jobs
        type: integer
      graphNodeId:
        description: GraphNodeID is the node ID in the graph (if this scenario run
          is part of a graph run)
        type: string
      graphRunName:
        description: GraphRunName is the name of the parent KrknGraphRun (if this
          scenario run is part of a graph run)
        type: string
      ownerUserId:
        description: OwnerUserID is the email address of the user who created this
          scenario run
        type: string
      phase:
        description: Phase is the overall phase of the scenario run
        type: string
      resiliencyScore:
        description: ResiliencyScore is the individual resiliency score for this scenario
          run node
        type: number
      resiliencyScoreEnabled:
        description: ResiliencyScoreEnabled indicates whether resiliency scoring is
          enabled for this run
        type: boolean
      resiliencyScores:
        description: ResiliencyScores contains per-cluster resiliency scores
        items:
          $ref: '#/definitions/internal_api.ClusterResiliencyScoreResponse'
        type: array
      runningJobs:
        description: RunningJobs is the number of currently running jobs
        type: integer
      scenarioImage:
        description: ScenarioImage is the resolved container image used by the scenario
          jobs
        type: string
      scenarioName:
        description: ScenarioName is the name of the scenario being executed
        type: string
      scenarioRunName:
        description: ScenarioRunName is the name of the KrknScenarioRun CR
        type: string
      successfulJobs:
        description: SuccessfulJobs is the number of successfully completed jobs
        type: integer
      totalTargets:
        description: TotalTargets is the total number of target clusters
        type: integer
    type: object
  internal_api.ScenarioRunListResponse:
    properties:
      pagination:
        allOf:
        - $ref: '#/definitions/internal_api.PaginationMeta'
        description: Pagination contains pagination metadata
      scenarioRuns:
        description: ScenarioRuns is the list of scenario runs
        items:
          $ref: '#/definitions/internal_api.ScenarioRunListItem'
        type: array
    type: object
  internal_api.ScenarioRunRequest:
    properties:
      categories:
        description: Categories are visible category names to associate with the run
          when it is created.
        items:
          type: string
        type: array
      cloudCredentialRef:
        description: |-
          CloudCredentialRef, if set, names a saved cloud credential Secret whose
          reference is set on the CRD spec for controller-level SecretKeyRef injection.
        type: string
      customRunName:
        description: CustomRunName is a user-provided label for the run (optional)
        type: string
      elasticsearchConfigName:
        description: |-
          ElasticsearchConfigName, if set, names a saved Elasticsearch config Secret whose
          credentials (ES_PASSWORD, and any ES_* vars not already in Environment) are
          injected server-side so the password is never transmitted by the client.
        type: string
      environment:
        additionalProperties:
          type: string
        description: Environment is a map of environment variables to pass to the
          container (optional)
        type: object
      fileReferences:
        description: FileReferences are references to centrally-managed files by UUID
          (optional)
        items:
          $ref: '#/definitions/github_com_krkn-chaos_krkn-operator_pkg_files.FileReference'
        type: array
      files:
        description: Files is an array of file objects to mount in the container (optional,
          legacy inline file mount)
        items:
          $ref: '#/definitions/internal_api.FileMount'
        type: array
      kubeconfigPath:
        description: 'KubeconfigPath is the path where kubeconfig should be mounted
          (optional, default: /home/krkn/.kube/config)'
        type: string
      registryName:
        description: |-
          RegistryName is the name of a saved registry (optional)
          If omitted, defaults to quay.io public registry
        type: string
      resiliencyScoreEnabled:
        description: |-
          ResiliencyScoreEnabled enables resiliency score calculation for this scenario run.
          When true, RESILIENCY_SCORE=true is set in scenario pods and the controller
          calculates scores from pod logs upon completion.
        type: boolean
      scenario:
        allOf:
        - $ref: '#/definitions/github_com_krkn-chaos_krkn-operator_api_v1alpha1.ScenarioReference'
        description: |-
          Scenario identifies the scenario and registry to resolve. The operator
          resolves the executable image through krknctl; callers cannot provide one.
      scenarioImage:
        description: |-
          ScenarioImage is accepted for backwards compatibility but is ignored.
          Images are always resolved server-side from Scenario.
          ScenarioImage is always serialized for config responses. It is empty
          until the controller has resolved an executable image for a job.
        type: string
      scenarioName:
        description: |-
          ScenarioName is the legacy scenario identity field. It is translated to
          Scenario when a request does not include the new reference object.
        type: string
      targetClusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: |-
          TargetClusters is a map of provider-name to list of cluster names
          Example: {"krkn-operator": ["cluster1", "cluster2"], "krkn-operator-acm": ["cluster3"]}
        type: object
      targetRequestId:
        description: TargetRequestID is the UUID of the KrknTargetRequest (required)
        type: string
    type: object
...
  internal_api_v2_websocket.ErrorMessage:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  internal_api_v2_websocket.ServerMessage:
    properties:
      data:
        description: payload
      event:
        description: '"updated", "deleted", "snapshot"'
        type: string
      id:
        description: resource ID (empty for dashboard/jobs)
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/internal_api_v2_websocket.WSPaginationMeta'
        description: pagination metadata (for "jobs" resource)
      resource:
        description: '"run", "graphrun", "dashboard", "jobs"'
        type: string
      stats:
        allOf:
        - $ref: '#/definitions/internal_api_v2_websocket.WSJobStatsSummary'
        description: aggregate job stats (for "jobs" resource)
      subscriptionId:
        description: request correlation for jobs snapshots
        type: string
    type: object
  internal_api_v2_websocket.WSJobStatsSummary:
    properties:
      failedJobs:
        type: integer
      succeededJobs:
        type: integer
      totalJobs:
        type: integer
    type: object
  internal_api_v2_websocket.WSPaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: krkn-chaos@googlegroups.com
    name: Krkn Team
    url: https://github.com/krkn-chaos/krkn-operator
  description: |-
    REST and WebSocket API for Krkn chaos engineering operator.

    **API Versions:**
    - **v1** - REST API with polling (deprecated but maintained)
    - **v2** - REST API with v2-specific resources + WebSocket real-time updates

    **WebSocket Authentication (v2):**
    WebSocket endpoints use JWT via subprotocol header:
    - JavaScript: `new WebSocket(url, 'access_token.' + jwtToken)`
    - Header: `Sec-WebSocket-Protocol: access_token.<jwt_token>`

    **Migration Path:**
    1. v1 REST → v2 REST (no changes, just update base path)
    2. v2 REST → v2 WebSocket (replace polling with multiplexed WebSocket)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://krkn-chaos.dev/terms
  title: Krkn Operator API
  version: "2.0"
paths:
...
  /graphruns/{graphRunName}/config:
    get:
      description: Retrieve graph run configuration directly by KrknGraphRun CR name
      parameters:
      - description: KrknGraphRun CR name
        in: path
        name: graphRunName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Graph run configuration
          schema:
            $ref: '#/definitions/internal_api.GraphRunCreateRequest'
        "400":
          description: Missing graph run name
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: User authentication required
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: GraphRun not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get graph run configuration
      tags:
      - graphruns
...
  /scenarios/run:
    get:
      description: Get list of all scenario runs with optional filtering by phase
        or scenario name
      parameters:
      - description: Filter by phase (Running, Succeeded, Failed)
        in: query
        name: phase
        type: string
      - description: Filter by scenario name
        in: query
        name: scenarioName
        type: string
      - description: Page number (1-based). Omit for all results.
        in: query
        name: page
        type: integer
      - description: Items per page (defaults to jobs.defaultPageSize config, fallback
          20; max 500). Only used when page is set.
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of scenario runs with pagination
          schema:
            $ref: '#/definitions/internal_api.ScenarioRunListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List scenario runs
      tags:
      - scenarios
    post:
      consumes:
      - application/json
      description: Execute a chaos scenario on target clusters with specified configuration
      parameters:
      - description: Scenario run configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.ScenarioRunRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Scenario run created
          schema:
            type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Target or scenario not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Run chaos scenario
      tags:
      - scenarios
  /scenarios/run/{jobID}:
    delete:
      description: Stop and delete a running or completed scenario run
      parameters:
      - description: Scenario run job ID
        in: path
        name: jobID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scenario run deleted
          schema:
            type: object
        "400":
          description: Invalid job ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Scenario run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete scenario run
      tags:
      - scenarios
...
  /scenarios/run/{scenarioRunName}:
    get:
      description: Get current execution status and metrics for a running or completed
        scenario
      parameters:
      - description: Scenario run name
        in: path
        name: scenarioRunName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scenario run status
          schema:
            type: object
        "400":
          description: Invalid scenario run name
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Scenario run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get scenario run status
      tags:
      - scenarios
  /scenarios/run/{scenarioRunName}/config:
    get:
      description: Retrieve scenario configuration directly by KrknScenarioRun CR
        name
      parameters:
      - description: KrknScenarioRun CR name
        in: path
        name: scenarioRunName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scenario configuration
          schema:
            $ref: '#/definitions/internal_api.ScenarioRunRequest'
        "400":
          description: Missing scenario run name
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: User authentication required
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: ScenarioRun not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get scenario run configuration
      tags:
      - scenarios
...
  /v2/ws/runs:
    get:
      consumes:
      - application/json
      description: |-
        Multiplexed WebSocket for real-time updates across all resources

        **Authentication:** JWT token via Sec-WebSocket-Protocol subprotocol
        - JavaScript: `new WebSocket(url, 'access_token.' + jwtToken)`
        - Header: `Sec-WebSocket-Protocol: access_token.<jwt_token>`

        **Client → Server Messages (subscribe/unsubscribe):**
        ```json
        {
        "action": "subscribe",
        "resource": "run",
        "ids": ["run-abc123", "run-xyz789"]
        }
        ```

        **Resource types:** `run`, `graphrun`, `dashboard`

        **Server → Client Messages (updates):**
        ```json
        {
        "resource": "run",
        "id": "run-abc123",
        "event": "updated",
        "data": { ... }
        }
        ```

        **Available endpoints:**
        - `/api/v2/ws/runs` - Subscribe to scenario run updates
        - `/api/v2/ws/graphruns` - Subscribe to graph run updates
        - `/api/v2/ws/jobs` - Subscribe to unified jobs list (paginated scenario runs + graph runs)
        - `/api/v2/ws/dashboard/active-runs` - Subscribe to dashboard updates
      produces:
      - application/json
      responses:
        "101":
          description: Switching protocols - WebSocket upgrade successful
          schema:
            $ref: '#/definitions/internal_api_v2_websocket.ServerMessage'
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            $ref: '#/definitions/internal_api_v2_websocket.ErrorMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api_v2_websocket.ErrorMessage'
      security:
      - BearerAuth: []
      summary: WebSocket real-time updates
      tags:
      - websocket
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /api/v1/auth/login or /api/v2/auth/login.
      Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"